Move followings to internal API
