+++ app/views/followings/organizations.json.jbuilder
